major cleanup in plot scripts and some restructuring
