+++ plot/archived_scripts/animate_plotly.py
