+++ plot/archived_scripts/create_single_image.py
