+++ plot/archived_scripts/field_plotter.py
